modulemd_tools: add the first pieces of a python library
Since the beginning of this project (in fact years before that) we
were always talking about creating a python library that would all our
tools use and share code with each other. A library that also others
(such as build-systems or other developers) could use.

I was hesitant to implement it because I was afraid that it will end
up reimplementing the whole `python3-libmodulemd` library (which would
be no good) or that we will create some nice, object-oriented and
high-level API, that would not fit some corner-cases and therefore
would be useless, because other people will have to implement it once
again but differently. I am finaly confident to add the first pieces
of the API.

As I see, it, there are three abstraction layers, that can be
implemented:

1. Functions, that would take modulemd YAMLs as a string, do something
   and return a new modulemd YAML string as their output. These can be
   easily implemented as pure functions with some small exceptions.
2. Functions, that would take and return objects from
   `python3-libmodulemd`
3. Collection of classes for representing modules, packages, etc and
   providing a facade over `python3-libmodule` objects, that would be
   used just internally and user will not have any knowing that they
   are used.

If requested or needed, I am going to implement all three layers,
separated that everyone could use the best approach for their
task. Now I am picking one, that I think will cover the most use-cases
with the least possible complexity, and that is (1).

I believe that this package will allow us to throw away a lot of
modularity-related code from Copr and change it to one-liners.

Lastly, the fact that I spent implementing this simple set of
functions almost a whole work week, shall be taken as a proof that
we really need such library. Others would have to do the same when
implementing modularity support for other tools / build systems, and
that would be a lot of time wasted.