createrepo_mod: use just createrepo_c if it has built-in module support
There is a built-in support for module metadata in createrepo_c since
version 0.16.1, please see the change log:

    rpm -q --changelog createrepo_c |less

In such case we can just call `createrepo_c` and be done with it. When
using older version of `createrepo_c` we still need to do the rest of
the process and call `modifyrepo_c`.