Loosen the python3-libmodulemd dependency to just libmodulemd
There is no `python3-libmodulemd` package for EPEL8, it is
a part of the `libmodulemd` package. For Fedora, we have both
options. So let's not if-else the spec and just install `libmodulemd`
everywhere and presume, that the python library will be available.

This is meant to be only a temporary solution until
`python3-libmodulemd` is available even for EPEL8, please see
https://bugzilla.redhat.com/show_bug.cgi?id=1860401

By depending on `libmodulemd` instead of `python3-libmodulemd` we
won't get any overrides to `libmodulemd`, which may cause
problems. Please see the discussion in PR #16.