modulemd_add_platform: Add --skip option to ignore modules which cannot support new platform
This is introduces a soft mode. Formerly, a modulemd-v2 document or
a modulemd-packager-v3 document without a context for the old platform
were handled as an error.

But when using this tool an mass where some people have not yet
migrated to modulemd-packager-v3 format or where people migrated but
they do want to build for the (previous) latest platform the hard mode
would not be helpfull.

This patch handles that cases as a non-error. Documents to log and
skip.