modulemd_add_platform: Don't crash on a document without a module name and with a default profile
convert_to_index() called for comparing serialized indices segfaults
if a modulemd-packager-v3 document has no module name, but it has
a profile marked as default. libmodulemd then attempts to generate
a modulemd-default-v1 document which requires a module name. If the
name is not set, an assertion fails and the library terminates:

    (process:41508): libmodulemd-CRITICAL **: 14:54:41.578: modulemd_defaults_set_module_name: assertion 'module_name' failed

    (process:41508): libmodulemd-CRITICAL **: 14:54:41.578: modulemd_defaults_v1_add_default_profile_for_stream: assertion 'stream_name' failed
    Neoprávněný přístup do paměti (SIGSEGV) (core dumped [obraz paměti uložen])

For some reason the crash does not occur without a stream name. Though
it should (libmodulemd-2.14.0). According to libmodulemd
documentation, an error instead of crash should be returned.

Nevertheless, modulemd_add_platform needs to adjust not to trigger the
error (once libmodulemd is fixed), the the crash (now).